getting ready for server work
Added functionality for ConversationTree to build itself from its
toString method. This provides an easy way to build a Conversation from
a string sent by offsite server. Also moved string to the top as public
static final variables. Might shove this into MyDatabase and other
places where the strings are used so it is easy to change structure
later.